+++ examples/interrupt_i2c.rs
@@ -6,6 +6,7 @@
 #![no_std]
 
 use core::sync::atomic::{AtomicUsize, Ordering};
+
 use defmt_rtt as _; // global logger
 
 use panic_probe as _;
@@ -38,10 +39,10 @@ dispatchers = [SPI1]
 mod app {
     use defmt::unwrap;
     use icm20948_driver::icm20948;
-    use stm32h7xx_hal::gpio::{self, Edge, Output, Input, PushPull, ExtiPin};
-    use stm32h7xx_hal::pac::SPI1;
+    use stm32h7xx_hal::gpio::{self, Output, Input, Edge, PushPull, ExtiPin};
+    use stm32h7xx_hal::i2c;
+    use stm32h7xx_hal::pac::I2C1;
     use stm32h7xx_hal::prelude::*;
-    use stm32h7xx_hal::spi;
     use systick_monotonic::{fugit::Duration, Systick};
 
     pub const MONO_TICK_RATE: u32 = 100;
@@ -59,7 +60,7 @@ mod app {
     struct Local {
         led: gpio::PE1<Output<PushPull>>,
         state: bool,
-        imu: icm20948::spi::IcmImu<spi::Spi<SPI1, spi::Enabled>, gpio::PD15<Output>>,
+        imu: icm20948::i2c::IcmImu<i2c::I2c<I2C1>>,
         int_pin: gpio::PE3<Input>,
     }
 
@@ -85,7 +86,7 @@ mod app {
             .pll1_q_ck(48.MHz())
             .freeze(pwrcfg, &syscfg);
 
-        // Setup heartbeat LED & interrupt pin
+        // Configure heartbeat LED and interrupt pin
         let gpioe = device.GPIOE.split(ccdr.peripheral.GPIOE);
         let led = gpioe.pe1.into_push_pull_output();
         let mut int_pin = gpioe.pe3.into_pull_up_input();
@@ -93,27 +94,16 @@ mod app {
         int_pin.trigger_on_edge(&mut exti, Edge::Rising);
         int_pin.enable_interrupt(&mut exti);
 
-
-        // Configure the SPI bus
-        let gpioa = device.GPIOA.split(ccdr.peripheral.GPIOA);
         let gpiob = device.GPIOB.split(ccdr.peripheral.GPIOB);
-        let gpiod = device.GPIOD.split(ccdr.peripheral.GPIOD);
+        let sda = gpiob.pb9.into_alternate_open_drain();
+        let scl = gpiob.pb8.into_alternate_open_drain();
 
-        let sck = gpioa.pa5.into_alternate();
-        let miso = gpioa.pa6.into_alternate();
-        let mosi = gpiob.pb5.into_alternate();
-        let mut cs = gpiod.pd15.into_push_pull_output();
-        cs.set_high();
+        let i2c = device
+            .I2C1
+            .i2c((scl, sda), 300.kHz(), ccdr.peripheral.I2C1, &ccdr.clocks);
+        // Configure i2c
 
-        let spi1: spi::Spi<_, _, u8> = device.SPI1.spi(
-            (sck, miso, mosi),
-            spi::MODE_0,
-            3.MHz(),
-            ccdr.peripheral.SPI1,
-            &ccdr.clocks,
-        );
-
-        let mut imu = unwrap!(icm20948::spi::IcmImu::new(spi1, cs));
+        let mut imu = unwrap!(icm20948::i2c::IcmImu::new(i2c, 0x68));
 
         unwrap!(imu.reset());
         defmt::debug!("IMU reset!");
@@ -121,8 +111,6 @@ mod app {
         unwrap!(imu.config_gyro_rate_div(5));              // While gyro is enabled this determines the ODR (output data rate). See data sheet.
         unwrap!(imu.config_acc_rate_div(5));
         unwrap!(imu.enable_int());                         // Enable the interrupt
-
-        cortex_m::asm::delay(SYS_TICK_RATE/2);
 
         heartbeat::spawn_after(Duration::<u64, 1, MONO_TICK_RATE>::from_ticks(
             MONO_TICK_RATE.into(),
@@ -175,7 +163,7 @@ mod app {
             acc[2]
         );
 
-       let gyr = unwrap!(cx.local.imu.read_gyro());
+        let gyr = unwrap!(cx.local.imu.read_gyro());
 
         defmt::debug!(
             "Gyro readings (dps): X: {}, Y: {}, Z: {}",