+++ src/icm20948.rs
@@ -126,7 +126,10 @@ pub enum GyroLPF {
 /// This may be caused by a number of reasons. For example, using the wrong 7-bit address with the I2C bus will cause a bus error.
 ///
 /// `InvalidInput` is for when an input to a driver function is unacceptable.
+#[derive(Debug)]
 #[derive(PartialEq)]
+#[derive(Copy)]
+#[derive(Clone)]
 pub enum IcmError<E> {
     /// An error occurred when using the bus
     BusError(E),
@@ -178,11 +181,11 @@ impl<E> From<E> for IcmError<E> {
     }
 }
 
-// TODO: Update with specific error
-impl<E> Format for IcmError<E> {
+#[cfg(feature = "defmt")]
+impl<E: Format> Format for IcmError<E> {
     fn format(&self, fmt: Formatter) {
-        match *self {
-            IcmError::BusError(_) => defmt::write!(fmt, "Bus Error!"),
+        match &*self {
+            IcmError::BusError(e) => defmt::write!(fmt, "Bus Error: {}!", e),
             IcmError::InvalidInput => defmt::write!(fmt, "Invalid input in the function!"),
         }
     }